added DishTools.java

+++ src/Dish.java
@@ -1,10 +1,10 @@
 public class Dish {
-    public double costInCents;
+    public int costInCents;
     public String nameOfDish;
     public boolean wouldRecommend;
 
     public void printSummary() {
-        System.out.printf("\nCost: %.0f cents\n Name: %s\n Recommended: %b" , costInCents, nameOfDish, wouldRecommend);
+        System.out.printf("\nCost: %d cents\n Name: %s\n Recommended: %b" , costInCents, nameOfDish, wouldRecommend);
     }
 
 }
